Allow specification of fps in floating point.
git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@1162 542714f4-19e9-0310-aa3c-eee0fc999fb1

+++ src/java/com/threerings/media/sprite/ImageSprite.java
@@ -1,5 +1,5 @@
 //
-// $Id: ImageSprite.java,v 1.1 2002/03/16 03:15:05 shaper Exp $
+// $Id: ImageSprite.java,v 1.2 2002/03/27 21:51:15 mdb Exp $
 
 package com.threerings.media.sprite;
 
@@ -141,9 +141,9 @@ public class ImageSprite extends Sprite
      *
      * @param fps the desired frames per second.
      */
-    public void setFrameRate (int fps)
+    public void setFrameRate (float fps)
     {
-        _frameDelay = 1000L/fps;
+        _frameDelay = (long)(1000/fps);
     }
 
     /**
